i965: Switch to roundf in HS/DS URB code.
Matt intentionally switched the VS calculation to be float-based in
commit c1da15709a0c0c2775bd9e534f67c60f7dc95ce8.  Tessellation support
was written before this and rebased forward, and missed the change.

Now it's consistent.

+++ src/mesa/drivers/dri/i965/gen7_urb.c
@@ -356,7 +356,7 @@ gen7_upload_urb(struct brw_context *brw, unsigned vs_size,
 
       if (total_wants > 0) {
          unsigned hs_additional = (unsigned)
-            round(hs_wants * (((double) remaining_space) / total_wants));
+            roundf(hs_wants * (((float) remaining_space) / total_wants));
          hs_chunks += hs_additional;
          remaining_space -= hs_additional;
          total_wants -= hs_wants;
@@ -364,7 +364,7 @@ gen7_upload_urb(struct brw_context *brw, unsigned vs_size,
 
       if (total_wants > 0) {
          unsigned ds_additional = (unsigned)
-            round(ds_wants * (((double) remaining_space) / total_wants));
+            roundf(ds_wants * (((float) remaining_space) / total_wants));
          ds_chunks += ds_additional;
          remaining_space -= ds_additional;
          total_wants -= ds_wants;
